plataforma_jota/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3-1.37.18
celery==5.3.6
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==2.1.0
Django==5.1.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.1
django-permissions-policy==4.18.0
django-prometheus==2.3.1
django_csp==3.8
djangorestframework==3.15.2
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==37.0.2
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
joblib==1.4.2
kombu==5.5.0
nltk==3.9.1
numpy==1.26.4
packaging==24.2
pika==1.3.2
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.9
pytest==8.0.2
pytest-django==4.8.0
python-dateutil==2.8.2
pytz==2025.1
PyYAML==6.0.1
regex==2024.11.6
requests==2.32.2
scikit-learn==1.5.1
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
